frontend-sfe: Refactor FAQ component
